pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "numpy>=2.3.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=21.0.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "scikit-learn>=1.7.1",
    "sqlalchemy>=2.0.41",
]
